pyproject.toml
[project]
name = "agentorg"
version = "0.1.0"
description = "Build your own AI organization — teams of agent roles that learn and grow."
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
fleet = "agentorg.cli.main:fleet"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-cov>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agentorg"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
